requirements.txt
cryptography==38.0.4
fastapi==0.115.12
httpx==0.28.1
python-dotenv==1.1.0
SQLAlchemy==2.0.23
uvicorn==0.34.3
jinja2==3.1.2
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib==1.7.4